move helper functions to utils
